dapp-update: reorganized /types and /utils folders
