+++ system-contract-dapp-playground/src/api/hedera/erc20-interactions/index.ts
@@ -19,7 +19,6 @@
  */
 
 import { Contract, isAddress } from 'ethers';
-import { ERC20MockSmartContractResult } from '@/types/interfaces';
 
 /**
  * @dev get token information
