+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-20/methods/TokenPermissions.tsx
@@ -43,11 +43,11 @@ import {
   useToast,
 } from '@chakra-ui/react';
 import {
-  allowanceParamFields,
   approveParamFields,
-  decreaseAllowanceParamFields,
   increaseAllowanceParamFields,
-} from '../../utils/constant';
+  decreaseAllowanceParamFields,
+  allowanceParamFields,
+} from '@/utils/contract-interactions/erc/constant';
 
 interface PageProps {
   baseContract: Contract;
