+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-20/methods/Transfer.tsx
@@ -25,7 +25,10 @@ import MultiLineMethod from '../common/MultiLineMethod';
 import { erc20Transfers } from '@/api/hedera/erc20-interactions';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { Dispatch, SetStateAction, useEffect, useState } from 'react';
-import { transferParamFields, transferFromParamFields } from '../../utils/constant';
+import {
+  transferParamFields,
+  transferFromParamFields,
+} from '@/utils/contract-interactions/erc/constant';
 
 interface PageProps {
   baseContract: Contract;
