+++ system-contract-dapp-playground/src/components/contract-interaction/index.tsx
@@ -23,16 +23,16 @@ import Image from 'next/image';
 import { Contract } from 'ethers';
 import ERC20Methods from './erc/erc-20/methods';
 import { deploySmartContract } from '@/api/hedera';
-import { HASHSCAN_BASE_URL } from '@/utils/constants';
 import HederaAlertDialog from '../common/AlertDialog';
 import { useCallback, useEffect, useState } from 'react';
 import { generateBaseContractInstance } from '@/api/ethers';
+import { HASHSCAN_BASE_URL } from '@/utils/common/constants';
 import ERC20DeployField from './erc/deployment/ERCDeployField';
-import { convertCalmelCaseFunctionName } from '@/utils/helpers';
 import { getHederaNativeIDFromEvmAddress } from '@/api/mirror-node';
-import { HederaContractAsset, NetworkName } from '@/types/interfaces';
+import { HederaContractAsset, NetworkName } from '@/types/common';
 import { CommonErrorToast, NoWalletToast } from '../toast/CommonToast';
 import { getInfoFromCookies, storeInfoInCookies } from '@/api/cookies';
+import { convertCalmelCaseFunctionName } from '@/utils/common/helpers';
 import { Tabs, TabList, Tab, TabPanels, useToast, TabPanel } from '@chakra-ui/react';
 import ExchangeRateDeployField from './exchange-rate-hip-475/deployment/ExchangeRateDeployField';
