+++ system-contract-dapp-playground/src/components/wallet-popup/index.tsx
@@ -22,8 +22,8 @@ import Link from 'next/link';
 import Image from 'next/image';
 import { ethers } from 'ethers';
 import { clearCookies } from '@/api/cookies';
-import { NetworkName } from '@/types/interfaces';
-import { HASHSCAN_BASE_URL } from '@/utils/constants';
+import { NetworkName } from '@/types/common';
+import { HASHSCAN_BASE_URL } from '@/utils/common/constants';
 import { BiCopy, BiCheckDouble } from 'react-icons/bi';
 import { SkeletonText, useToast } from '@chakra-ui/react';
 import { getBalance, getWalletProvider } from '@/api/wallet';
