+++ system-contract-dapp-playground/src/sections/hrc-719/index.tsx
@@ -23,7 +23,7 @@
 import Link from 'next/link';
 import { motion } from 'framer-motion';
 import { IoOpenOutline } from 'react-icons/io5';
-import { HEDERA_OFFICIAL_HIPS_URL, HEDERA_SMART_CONTRACTS_ASSETS } from '@/utils/constants';
+import { HEDERA_OFFICIAL_HIPS_URL, HEDERA_SMART_CONTRACTS_ASSETS } from '@/utils/common/constants';
 import { Tooltip } from '@chakra-ui/react';
 import ContractInteraction from '@/components/contract-interaction';
 import { BsGithub } from 'react-icons/bs';
