+++ system-contract-dapp-playground/src/sections/overview/index.tsx
@@ -22,7 +22,7 @@
 
 import Link from 'next/link';
 import { motion } from 'framer-motion';
-import { LEFT_SIDE_BAR_ITEMS } from '@/utils/constants';
+import { LEFT_SIDE_BAR_ITEMS } from '@/utils/common/constants';
 
 const OverviewSection = () => {
   return (
