+++ system-contract-dapp-playground/src/sections/prng-hip-351/index.tsx
@@ -26,7 +26,7 @@ import { BsGithub } from 'react-icons/bs';
 import { Tooltip } from '@chakra-ui/react';
 import { IoOpenOutline } from 'react-icons/io5';
 import ContractInteraction from '@/components/contract-interaction';
-import { HEDERA_OFFICIAL_HIPS_URL, HEDERA_SMART_CONTRACTS_ASSETS } from '@/utils/constants';
+import { HEDERA_OFFICIAL_HIPS_URL, HEDERA_SMART_CONTRACTS_ASSETS } from '@/utils/common/constants';
 
 const HIP351Section = () => {
   return (
