+++ system-contract-dapp-playground/src/utils/common/metadata.ts
